+++ source/s2/base/panic.h
@@ -1,7 +1,9 @@
 #pragma once
 
+#include "s2/base/basic_types.h"
+
 #define s2_panic(msg) ::s2::base::internal::panic(msg, __FILE__, __LINE__)
 
 namespace s2::base::internal {
-[[noreturn]] void panic(char const* msg, char const* file, unsigned int line);
+[[noreturn]] void panic(char const* msg, char const* file, usize line);
 }
